requirements.txt
django>=6.0
djangorestframework>=3.17
django-cors-headers>=4.9
celery>=5.6
redis>=7.4
psycopg2-binary>=2.9
python-decouple>=3.8
pydantic>=2.12
requests>=2.32
gunicorn>=23.0
